Server 0.0.2 additions
Using naming convention, renamed a lot of commands and functions used within the server console, e.g. deleteFolder named to folderDelete, so "big-to-small" is adhered to. Added returns of true/false, and error throws instead of returning ambiguous errors which should not be standard because of the parsing it involves. Added server console commands to be added next.