Added presetResetElements and presetAddElements
Quickly added presetResetElement function (because of necessity it was pretty annoying having to manually reset the elemetns of a perfab), and prefabAddElement I was working on already: allows you to add an element object to a prefab - does some simple, not too clever, name duplication check so it doesn't add 2 of the same elements names in the same prefab. does the job.